Refactor: Moved PhysicalTable implementations to their own package.
