+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/table/DimensionListPartitionTableDefinition.java
@@ -10,7 +10,7 @@ import com.yahoo.bard.webservice.data.dimension.Dimension;
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
-import com.yahoo.bard.webservice.table.BaseCompositePhysicalTable;
+import com.yahoo.bard.webservice.table.physicaltables.BaseCompositePhysicalTable;
 import com.yahoo.bard.webservice.table.ConfigPhysicalTable;
 import com.yahoo.bard.webservice.table.availability.PartitionAvailability;
 import com.yahoo.bard.webservice.table.resolver.DataSourceFilter;
