+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/physicaltables/StrictPhysicalTable.java
@@ -1,11 +1,12 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.table;
+package com.yahoo.bard.webservice.table.physicaltables;
 
 import com.yahoo.bard.webservice.data.config.names.DataSourceName;
 import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
+import com.yahoo.bard.webservice.table.Column;
 import com.yahoo.bard.webservice.table.availability.StrictAvailability;
 
 import org.joda.time.DateTime;
