+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/config/table/BaseTableLoadTaskSpec.groovy
@@ -19,7 +19,7 @@ import com.yahoo.bard.webservice.data.time.DefaultTimeGrain
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
 import com.yahoo.bard.webservice.table.ConfigPhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTableSchema
-import com.yahoo.bard.webservice.table.StrictPhysicalTable
+import com.yahoo.bard.webservice.table.physicaltables.StrictPhysicalTable
 import com.yahoo.bard.webservice.table.TableGroup
 
 import org.joda.time.DateTimeZone
