+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/config/table/DimensionListPartitionTableDefinitionSpec.groovy
@@ -10,7 +10,7 @@ import com.yahoo.bard.webservice.data.dimension.Dimension
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary
 import com.yahoo.bard.webservice.data.dimension.DimensionField
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
-import com.yahoo.bard.webservice.table.BaseCompositePhysicalTable
+import com.yahoo.bard.webservice.table.physicaltables.BaseCompositePhysicalTable
 import com.yahoo.bard.webservice.table.ConfigPhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTableSchema
